Add files via upload
The oblective of this project is to propose an architecture to develop a comprehensive data warehouse that provides basic database analytics and reporting services, gathers data from multiple sources, and creates metrics to drive clinical and operational performance improvement. The primary focus will be on the 1115 Medicaid Waiver DPP, while also supporting additional departmental projects.